Moved offset line into line class.
