+++ src/GShark.Test.XUnit/Operation/OffsetTests.cs
@@ -18,20 +18,6 @@ namespace GShark.Test.XUnit.Operation
         }
 
         [Fact]
-        public void Returns_The_Offset_Of_A_Line()
-        {
-            // Arrange
-            Line ln = new Line(new Point3(5, 0, 0), new Point3(0, 5, 0));
-            double offset = 12;
-
-            // Act
-            Line offsetResult = Offset.Line(ln, offset, Plane.PlaneXY);
-
-            // Assert
-            (offsetResult.StartPoint.DistanceTo(ln.StartPoint) - offset).Should().BeLessThan(GSharkMath.MaxTolerance);
-        }
-
-        [Fact]
         public void Returns_The_Offset_Of_A_Circle()
         {
             // Arrange
